requirements/dev.txt
# run like $ pip install -r requirements/dev.txt
-r prod.txt
django-debug-toolbar==1.2.2
